Servers: split error-reporting server from the lobby
