+++ http-clients/error-reporting-client/build.gradle
@@ -1,0 +1,10 @@
+dependencies {
+    implementation "io.github.openfeign:feign-core:$feignCoreVersion"
+    implementation "io.github.openfeign:feign-gson:$feignGsonVersion"
+    implementation project(":http-clients:feign-common")
+    implementation project(":http-clients:http-client-lib")
+    implementation project(":lib:java-extras")
+    testImplementation "ru.lanwen.wiremock:wiremock-junit5:$wireMockJunit5Version"
+    testImplementation "com.github.tomakehurst:wiremock:$wireMockVersion"
+    testImplementation project(":lib:test-common")
+}
