+++ http-clients/error-reporting-client/src/main/java/org/triplea/http/client/error/report/CanUploadRequest.java
@@ -1,0 +1,19 @@
+package org.triplea.http.client.error.report;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@ToString
+@EqualsAndHashCode
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class CanUploadRequest {
+  private String gameVersion;
+  private String errorTitle;
+}
