+++ http-clients/http-client-lib/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+    implementation project(":game-app:domain-data")
+}
