+++ servers/README-WIP.md
@@ -10,5 +10,3 @@ the code a bit to be yet more modular.
 Further, with very specific sub-folders and more executables, we will have more opportunity
 to do continuous deployments on any update without having to worry for example about
 disconnecting everyone from chat.
-
-
