+++ servers/error-reporting/configuration.yml
@@ -1,0 +1,61 @@
+githubApiToken: ${GITHUB_API_TOKEN:-}
+githubWebServiceUrl: https://api.github.com
+githubGameOrg: triplea-game
+githubGameRepo: triplea
+githubMapsOrgName: triplea-maps
+
+# When disabled, API calls to github to create error reports will not
+# be made and instead a hardcoded value returned.
+errorReportToGithubEnabled: ${ERROR_REPORT_TO_GITHUB_ENABLED:-false}
+
+# Whether to print out SQL statements as executed, useful for debugging.
+logSqlStatements: false
+
+database:
+  driverClass: org.postgresql.Driver
+  user: ${DATABASE_USER:-error_report_user}
+  password: ${DATABASE_PASSWORD:-error_report}
+  url: jdbc:postgresql://${DB_URL:-localhost:5432/error_report}
+  properties:
+    charSet: UTF-8
+
+  # the maximum amount of time to wait on an empty pool before throwing an exception
+  maxWaitForConnection: 1s
+
+  # the SQL query to run when validating a connection's liveness
+  validationQuery: select 1
+
+  # the minimum number of connections to keep open
+  minSize: 8
+
+  # the maximum number of connections to keep open
+  maxSize: 32
+
+  # whether or not idle connections should be validated
+  checkConnectionWhileIdle: false
+
+  # the amount of time to sleep between runs of the idle connection validation, abandoned cleaner and idle pool resizing
+  evictionInterval: 10s
+
+  # the minimum amount of time an connection must sit idle in the pool before it is eligible for eviction
+  minIdleTime: 1 minute
+
+
+logging:
+  # The default level of all loggers. Can be OFF, ERROR, WARN, INFO, DEBUG, TRACE, or ALL.
+  level: INFO
+  loggers:
+    # Set this to DEBUG to troubleshoot HTTP 400 "Unable to process JSON" errors.
+    io.dropwizard.jersey.jackson.JsonProcessingExceptionMapper: INFO
+
+server:
+  applicationConnectors:
+    - type: http
+      port: ${HTTP_PORT:-8090}
+      # useForwardedHeaders is important for when behind a reverse proxy (NGINX)
+      # Without this 'getRemoteAddr' will return the IP of the reverse proxy server.
+      # By default when building locally useForwardedPorts should be 'false', but
+      # for all other environments that do have a NGINX server, the value should be
+      # set to true.
+      useForwardedHeaders: ${USE_FORWARDED_HEADERS:-false}
+  adminConnectors: []
