+++ servers/error-reporting/src/main/java/org/triplea/server/error/reporting/ErrorReportingConfiguration.java
@@ -1,0 +1,25 @@
+package org.triplea.server.error.reporting;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import io.dropwizard.Configuration;
+import io.dropwizard.db.DataSourceFactory;
+import javax.validation.Valid;
+import javax.validation.constraints.NotNull;
+import lombok.Data;
+import lombok.Getter;
+
+@Data
+public class ErrorReportingConfiguration extends Configuration {
+
+  String githubApiToken;
+  String githubWebServiceUrl;
+
+  String githubGameOrg; // : triplea-game
+  String githubGameRepo; // triplea
+  String githubMapsOrgName; // : triplea-maps
+
+  Boolean errorReportToGithubEnabled;
+  Boolean logSqlStatements;
+
+  @Valid @NotNull @JsonProperty @Getter DataSourceFactory database = new DataSourceFactory();
+}
