+++ servers/error-reporting/src/test/resources/datasets/error_reporting/empty_error_report_history.yml
