+++ servers/error-reporting/src/test/resources/db-cleanup.sql
