+++ servers/server-lib/src/main/java/org/triplea/dropwizard/common/JdbiLogging.java
