+++ servers/server-test-support/build.gradle
@@ -1,0 +1,28 @@
+dependencies {
+//    implementation "com.liveperson:dropwizard-websockets:$dropwizardWebsocketsVersion"
+//    implementation "io.dropwizard:dropwizard-auth:$dropwizardVersion"
+//    implementation "io.dropwizard:dropwizard-core:$dropwizardVersion"
+//    implementation "io.dropwizard:dropwizard-jdbi3:$dropwizardVersion"
+//    implementation project(':game-app:domain-data')
+//    implementation project(':http-clients:feign-common')
+//    implementation project(':http-clients:github-client')
+//    implementation project(':http-clients:lobby-client')
+//    implementation project(':lib:java-extras')
+//    implementation project(':lib:websocket-client')
+//    implementation project(':lib:websocket-server')
+//    implementation project(':servers:server-lib')
+//    implementation project(':spitfire-server:latest-version-module')
+//    implementation project(':spitfire-server:lobby-module')
+//    implementation project(':spitfire-server:maps-module')
+
+//    testImplementation "com.github.database-rider:rider-junit5:$databaseRiderVersion"
+    implementation "io.dropwizard:dropwizard-testing:$dropwizardVersion"
+//    testImplementation "io.github.openfeign:feign-core:$feignCoreVersion"
+//    testImplementation "io.github.openfeign:feign-gson:$feignGsonVersion"
+//    testImplementation "org.awaitility:awaitility:$awaitilityVersion"
+    implementation "org.junit.jupiter:junit-jupiter-api:$junitJupiterVersion"
+//    testImplementation project(':spitfire-server:database-test-support')
+//    testImplementation project(':lib:test-common')
+//    testImplementation "org.java-websocket:Java-WebSocket:$javaWebSocketVersion"
+//    runtimeOnly "org.postgresql:postgresql:$postgresqlVersion"
+}
