+++ settings.gradle
@@ -7,8 +7,10 @@ include 'game-app:game-headless'
 include 'game-app:game-relay-server'
 include 'game-app:map-data'
 include 'game-app:smoke-testing'
+include 'http-clients:error-reporting-client'
 include 'http-clients:feign-common'
 include 'http-clients:github-client'
+include 'http-clients:http-client-lib'
 include 'http-clients:lobby-client'
 include 'lib:java-extras'
 include 'lib:swing-lib'
@@ -17,10 +19,12 @@ include 'lib:test-common'
 include 'lib:websocket-client'
 include 'lib:websocket-server'
 include 'lib:xml-reader'
+include 'servers:error-reporting'
+include 'servers:server-lib'
+include 'servers:server-test-support'
 include 'spitfire-server:database'
 include 'spitfire-server:database-test-support'
 include 'spitfire-server:dropwizard-server'
-include 'spitfire-server:error-reporting-module'
 include 'spitfire-server:latest-version-module'
 include 'spitfire-server:lib:scheduled-tasks'
 include 'spitfire-server:lobby-module'
