--- spitfire-server/dropwizard-server/src/main/java/org/triplea/dropwizard/common/AuthenticationConfiguration.java
@@ -1,49 +1,0 @@
-package org.triplea.dropwizard.common;
-
-import com.codahale.metrics.MetricRegistry;
-import com.github.benmanes.caffeine.cache.Caffeine;
-import io.dropwizard.auth.AuthDynamicFeature;
-import io.dropwizard.auth.AuthValueFactoryProvider;
-import io.dropwizard.auth.Authenticator;
-import io.dropwizard.auth.Authorizer;
-import io.dropwizard.auth.CachingAuthenticator;
-import io.dropwizard.auth.oauth.OAuthCredentialAuthFilter;
-import io.dropwizard.setup.Environment;
-import java.security.Principal;
-import java.time.Duration;
-import lombok.experimental.UtilityClass;
-import org.glassfish.jersey.server.filter.RolesAllowedDynamicFeature;
-import org.triplea.http.client.lobby.AuthenticationHeaders;
-
-@UtilityClass
-public class AuthenticationConfiguration {
-
-  /**
-   * Enables configuration via OAuth token. Endpoints annotated with @RolesAllowed will be activated
-   * and will require a user to have been a given role during per-request authentication.
-   */
-  public static <UserT extends Principal> void enableAuthentication(
-      final Environment environment,
-      final MetricRegistry metrics,
-      final Authenticator<String, UserT> authenticator,
-      final Authorizer<UserT> authorizer,
-      final Class<UserT> principalClass) {
-    environment
-        .jersey()
-        .register(
-            new AuthDynamicFeature(
-                new OAuthCredentialAuthFilter.Builder<UserT>()
-                    .setAuthenticator(
-                        new CachingAuthenticator<>(
-                            metrics,
-                            authenticator,
-                            Caffeine.newBuilder()
-                                .expireAfterAccess(Duration.ofMinutes(10))
-                                .maximumSize(10000)))
-                    .setAuthorizer(authorizer)
-                    .setPrefix(AuthenticationHeaders.KEY_BEARER_PREFIX)
-                    .buildAuthFilter()));
-    environment.jersey().register(new AuthValueFactoryProvider.Binder<>(principalClass));
-    environment.jersey().register(new RolesAllowedDynamicFeature());
-  }
-}
