--- spitfire-server/error-reporting-module/src/main/java/org/triplea/modules/error/reporting/CreateIssueParams.java
@@ -1,14 +1,0 @@
-package org.triplea.modules.error.reporting;
-
-import javax.annotation.Nonnull;
-import lombok.Builder;
-import lombok.Value;
-import org.triplea.http.client.error.report.ErrorReportRequest;
-
-@Value
-@Builder(toBuilder = true)
-public class CreateIssueParams {
-  @Nonnull private final String ip;
-  @Nonnull private final String systemId;
-  @Nonnull private final ErrorReportRequest errorReportRequest;
-}
